Gradescope Buuble Sheet Serial Number Automation
It is a python script that adds serial numbers to a gradescope bubble sheet. You can use this to print "n" number of Gradescope bubble sheets with serial numbers ranging from 1 to n. The only requirement is a valid gradescope bubble sheet as input. The numbers will be added in the "Other" section.